% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here_S.R
\name{here_S}
\alias{here_S}
\title{File path to S drive}
\usage{
here_S(
  file_path,
  name,
  folder = c("r", "rw"),
  server_path = "S:/010 SLOŽKY ZAMĚSTNANCU - FOLDERS OF EMPLOYEES"
)
}
\arguments{
\item{file_path}{character; path to file.}

\item{name}{character; name of employee, usually in the format "LASTNAME FIRSTNAME".}

\item{folder}{character; indicates whether to set the path to the "READ ONLY"
or the "READ WRITE FREE" folder.}

\item{server_path}{character; hard-coded path to the employee folders on S drive.}
}
\description{
Convenience function to set the file path to the private folder
on S drive.
}
