% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chnso.R
\name{read_chnso}
\alias{read_chnso}
\title{Import data from FLASH 2000}
\usage{
read_chnso(file_path, mode = "chns")
}
\arguments{
\item{file_path}{character; path to file}

\item{mode}{character; analysis mode. usually either "chns" or "o".}
}
\description{
This function reads in data from the Thermo Scientific FLASH 2000 CHNS-O
analyzer located at the Husova laboratory.
}
\author{
Anil Axel Tellbüscher
}
