% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ms3000.R
\name{read_ms3000}
\alias{read_ms3000}
\title{Imoprt data from MasterSizer 3000}
\usage{
read_ms3000(file_path, decimal_mark = ",")
}
\arguments{
\item{file_path}{character; path to file}

\item{decimal_mark}{character; indicating the decimal mark to be used. The
default is comma, due to the language settings of the computer.}
}
\value{
a tibble
}
\description{
Convenience function to import data files from Malvern Panalytics
MasterSizer 3000 laser diffraction particle analyzer equipped with
a HydroEV unit.
}
\details{
The device is located at the Husova building.
}
\author{
Anil Axel Tellbüscher
}
