% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{read_toc}
\alias{read_toc}
\title{Import data from Shimadzu TOC-L/TNM-L}
\usage{
read_toc(file_path, ..., header_keywords = "Data")
}
\arguments{
\item{file_path}{character; path to the file to load}

\item{...}{further arguments passed to or from other methods.}

\item{header_keywords}{character vector; keywords used to identify the header line
in the Shimadzu TOC-L/TNM-L output file. The function scans each line and selects
the first line that contains all specified keywords. This allows dynamic detection
of the data table start, even if the number of header lines varies.
Default is \code{c("Sample Name", "Result(TOC)")}.}
}
\value{
a tibble
}
\description{
Function that imports either a basic or a detailed data report file
 from the Shimadzu TOC-L/TNM-L analyzer located in the small lab at
 ZR.
}
\author{
Anil Axel Tellbüscher
}
