% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_toctn.R
\name{read_toctn}
\alias{read_toctn}
\title{Import data from Shimadzu TOC-L/TNM-L}
\usage{
read_toctn(file_path, detailed = FALSE, decimal_mark = ",")
}
\arguments{
\item{file_path}{character; path to the file to load}

\item{detailed}{logical; indicating whether the file to be loaded is a basic
report (only summarized results) or a detailed report (including data from
individual injections).}

\item{decimal_mark}{character; indicating the decimal mark to be used. The
default is comma, due to the language settings of the computer.}
}
\value{
a tibble
}
\description{
Function that imports either a basic or a detailed data report file
 from the Shimadzu TOC-L/TNM-L analyzer located in ZR 02 040.
}
\author{
Anil Axel Tellbüscher
}
