% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_functions.R
\name{rename_aas}
\alias{rename_aas}
\title{Rename data from Shimadzu F-AAS}
\usage{
rename_aas(data)
}
\arguments{
\item{data}{input data to rename columns}
}
\value{
a tibble
}
\description{
Function to automatically rename column names of data imports from the
Shimadzu AA-7000 Flame-Atomic Absorption Spectroscope (F-AAS).
}
\details{
Column name mappings used in the function:\cr
\itemize{
  \describe{type: }{Sample type:\cr
    - STD: Standard used to create the calibration curve\cr
    - BLK: Blank sample\cr
    - UNK: Unknown sample\cr
    Mapped from "Action"}

  \item{id_sample: }{Sample ID specified by the user.\cr
    Mapped from "Sample ID"}

  \item{X: }{Mapped from "X"}
  \item{M: }{Mapped from "M"}
  \item{Q: }{Mapped from "Q"}

  \item{conc_true: }{Concentration of the calibration standard specified by the user.\cr
    Values are reported in mg/L.\cr
    Mapped from "True Value (mg/L)"}

  \item{conc: }{Concentration calculated by WizAArd.\cr
    Values are reported in mg/L.\cr
    Mapped from "Conc. (mg/L)"}

  \item{abs: }{Measured absorbance value.\cr
    Mapped from "Abs."}

  \item{bg: }{Measured background signal emitted by the deuterium lamp.\cr
    Mapped from "BG"}

  \item{process: }{Mapped from "Data Process Range (sec)"}
  \item{position}{Mapped from "Pos."}
  \item{V: }{Mapped from "VOL"}
  \item{diluent: }{Mapped from "Diluent"}
  \item{reagent1: }{Mapped from "Reagent 1"}
  \item{reagent2: }{Mapped from "Reagent 2"}
  \item{reagent3: }{Mapped from "Reagent 3"}
  \item{V_total: }{Mapped from "Total Volume"}
  \item{factor_m: }{Mapped from "WF"}
  \item{factor_V: }{Mapped from "VF"}
  \item{factor_dil: }{Mapped from "DF"}
  \item{factor_dil_asc: }{Mapped from "ASC DF"}
  \item{factor_cor: }{Mapped from "CF"}

  \item{conc_actual: }{Recalculated concentration based on calculated concentration and all correction factors.\cr
    Mapped from "Actual Conc."}

  \item{unit_actual: }{Unit of the recalculated concentration.\cr
    Mapped from "Actual Conc. Unit"}

  \item{rsd: }{Mapped from "\%RSD"}
  \item{sd: }{Mapped from "SD"}
  \item{rpd: }{Mapped from "RPD"}
  \item{recovery: }{Mapped from "\%R"}

  \item{id_cal: }{Identifier of the calibration curve.\cr
    If set for an unknown sample, this indicates which calibration was used.\cr
    Mapped from "C#"}

  \item{sample_group: }{Mapped from "SG#"}
  \item{remark: }{Mapped from "Out of Control Remark"}
  \item{date: }{Mapped from "Date"}
  \item{time: }{Mapped from "Time"}
  \item{user: }{Mapped from "User Name"}
  \item{device: }{Mapped from "Device Name"}
}
}
\author{
Anil Axel Tellbüscher
}
