% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_functions.R
\name{rename_chnso}
\alias{rename_chnso}
\title{Rename data from ThermoFisher FLASH 2000}
\usage{
rename_chnso()
}
\description{
Rename data from ThermoFisher FLASH 2000
}
\author{
Anil Axel Tellbüscher
}
