% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_toctn.R
\name{rename_toctn}
\alias{rename_toctn}
\title{Rename data from Shimadzu TOC-L/TNM-L}
\usage{
rename_toctn(data)
}
\arguments{
\item{data}{input data to rename columns}
}
\value{
something
}
\description{
Function to automatically rename column names of data imports from the
Shimadzu TOC-L/TNM-L analyzer.
}
\author{
Anil Axel Tellbüscher
}
