% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_functions.R
\name{select_aas_subset}
\alias{select_aas_subset}
\title{Select variables from Shimadzu F-AAS data}
\usage{
select_aas_subset(data, subset = "flame")
}
\arguments{
\item{data}{A tibble or data frame that has already been renamed using `rename_aas()`.}

\item{subset}{A character string specifying the variable group to select.
Options are:
- `"all"` (default): select all renamed Shimadzu variables
- `"calibration"`: variables related to calibration standards and factors
- `"absorbance"`: absorbance and background signal
- `"sample_info"`: sample identifiers and metadata
- `"dilution"`: reagents and dilution details}
}
\value{
A tibble with the selected subset of columns
}
\description{
This function selects subsets of variables from a dataset that has been
renamed using `rename_aas()`. You can choose predefined subsets, such as
"calibration", "absorbance", "sample_info", etc.
}
