% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here_S.R
\name{here_S}
\alias{here_S}
\title{File path to S: drive}
\usage{
here_S(
  file_path,
  name = Sys.getenv("FFPW_NAME"),
  folder = "w",
  server_path = "S:/010 SLOŽKY ZAMĚSTNANCU - FOLDERS OF EMPLOYEES"
)
}
\arguments{
\item{file_path}{character; path to file.}

\item{name}{character; name of employee, usually in the format
"LASTNAME FIRSTNAME". The name is read from .Renviron by default.}

\item{folder}{character; indicates whether to set the path to the
"READ ONLY" or the "READ WRITE FREE" folder.}

\item{server_path}{character; hard-coded path to the employee folders
on S: drive.}
}
\description{
The `here_S()` function provides a convenient shortcut to access
the private folder on S: drive. Note that the function only works
when the computer is connected  to the university network.
}
