% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{read_aas}
\alias{read_aas}
\title{Import data from Shimadzu F-AAS}
\usage{
read_aas(file_path, skip_lines = 2, device_setup = "flame", ...)
}
\arguments{
\item{file_path}{character; path to the file to load.}

\item{skip_lines}{numeric; number of lines to be skipped when
reading the file. The value depends on how many lines the initial
comment at the beginning of a measurement file has.}

\item{device_setup}{character; default is "flame"}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a tibble.
}
\description{
Function that imports either a basic or a detailed data report file
 from the Shimadzu AA-7000 Flame-Atomic absorption spectroscope
 located in the small lab at ZR.
}
\author{
Anil Axel Tellbüscher
}
