% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{read_toc}
\alias{read_toc}
\title{Import data from Shimadzu TOC-L/TNM-L}
\usage{
read_toc(file_path, skip_lines = 9, ...)
}
\arguments{
\item{file_path}{character; path to the file to load}

\item{skip_lines}{numeric; indicating how many lines of the input file should be
skipped}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a tibble
}
\description{
Function that imports either a basic or a detailed data report file
 from the Shimadzu TOC-L/TNM-L analyzer located in the small lab at
 ZR.
}
\author{
Anil Axel Tellbüscher
}
