% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_aas.R
\name{rename_aas}
\alias{rename_aas}
\title{Rename data from Shimadzu F-AAS}
\usage{
rename_aas(data)
}
\arguments{
\item{data}{input data to rename columns}
}
\value{
something
}
\description{
Function to automatically rename column names of data imports from the
Shimadzu AA-7000 Flame-Atomic Absorption Spectroscope (F-AAS).
}
\author{
Anil Axel Tellbüscher
}
